chore: cleaner files
- restructure tests and function order in files
- added some jsdoc
