+++ src/cardDataExtraction.ts
@@ -1,5 +1,20 @@
+import {
+	isDefinedAsString,
+	isDefined,
+	isDefinedAsArray,
+	getMetadataKey,
+	parseAbstractDate,
+} from "~/utils";
+
+import { FrontMatterCache, TagCache, TFile } from "obsidian";
+import type {
+	AutoTimelineSettings,
+	AbstractDate,
+	MarkdownCodeBlockTimelineProcessingContext,
+} from "~/types";
+
 /**
- * Returns a list of tags to for in a note based of plugin settings, note frontmatter and note tags.
+ * Returns a list of tags based off plugin settings, note frontmatter and note tags.
  *
  * @param { AutoTimelineSettings } settings - The plugins settings.
  * @param { FrontMatterCache } metaData - The frontematter cache.
@@ -34,27 +49,105 @@ export function getTagsFromMetadataOrTagObject(
 	return output;
 }
 
-export function getAbstractDateFromData(
-	groupsToCheck: string[],
-	metadataString: string,
-	reg: RegExp | string
+/**
+ * Extract the body from the raw text of a note.
+ * @decsription After extraction most markdown tokens will be removed and links will be sanitized aswell and wrapped into bold tags for clearner display.
+ *
+ * @param { string } rawFileText - The text content of a obsidian note.
+ * @param { MarkdownCodeBlockTimelineProcessingContext } context - Timeline generic context.
+ * @returns { string | null } the body of a given card or null if none was found.
+ */
+export function getBodyFromContextOrDocument(
+	rawFileText: string,
+	context: MarkdownCodeBlockTimelineProcessingContext
+): string | null {
+	const {
+		cachedMetadata: { frontmatter: metadata },
+		settings: { metadataKeyEventBodyOverride },
+	} = context;
+	const overrideBody = metadata?.[metadataKeyEventBodyOverride] ?? null;
+
+	if (!rawFileText.length || overrideBody) return overrideBody;
+
+	const rawTextArray = rawFileText.split("\n");
+	rawTextArray.shift();
+	const processedArray = rawTextArray.slice(rawTextArray.indexOf("---") + 1);
+	const finalString = processedArray.join("\n").trim();
+
+	return finalString;
+}
+
+/**
+ * Extract the first image from the raw markdown in a note.
+ *
+ * @param { string } rawFileText - The text content of a obsidian note.
+ * @param { MarkdownCodeBlockTimelineProcessingContext } context - Timeline generic context.
+ * @returns { string | null } the URL of the image to be displayed in a card or null if none where found.
+ */
+export function getImageUrlFromContextOrDocument(
+	rawFileText: string,
+	context: MarkdownCodeBlockTimelineProcessingContext
+): string | null {
+	const {
+		cachedMetadata: { frontmatter: metadata },
+		file: currentFile,
+		app,
+		settings: { metadataKeyEventPictureOverride },
+	} = context;
+	const {
+		vault,
+		metadataCache: { getFirstLinkpathDest },
+	} = app;
+	const override = metadata?.[metadataKeyEventPictureOverride];
+
+	if (override) return override;
+	const internalLinkMatch = rawFileText.match(/!\[\[(?<src>.*)\]\]/);
+	const matchs =
+		internalLinkMatch || rawFileText.match(/!\[.*\]\((?<src>.*)\)/);
+
+	if (!matchs || !matchs.groups || !matchs.groups.src) return null;
+
+	if (internalLinkMatch) {
+		// https://github.com/obsidianmd/obsidian-releases/pull/1882#issuecomment-1512952295
+		const file = getFirstLinkpathDest.bind(app.metadataCache)(
+			matchs.groups.src,
+			currentFile.path
+		) as TFile | null;
+
+		if (file instanceof TFile) return vault.getResourcePath(file);
+		// Thanks https://github.com/joethei
+		return null;
+	} else return encodeURI(matchs.groups.src);
+}
+
+/**
+ * Given a metadata key it'll try to parse the associated data as an `AbstractDate` and return it
+ *
+ * @param { MarkdownCodeBlockTimelineProcessingContext } param0 - Timeline generic context.
+ * @param { string } key - The target lookup key in the notes metadata object.
+ * @returns { AbstractDate | undefined } the abstract date representation or undefined.
+ */
+export function getAbstractDateFromMetadata(
+	{ cachedMetadata, settings }: MarkdownCodeBlockTimelineProcessingContext,
+	key: string
 ): AbstractDate | undefined {
-	const matches = metadataString.match(reg);
+	const groupsToCheck = settings.dateParserGroupPriority.split(",");
+	const numberValue = getMetadataKey(cachedMetadata, key, "number");
 
-	if (!matches || !matches.groups) return undefined;
+	if (isDefined(numberValue)) {
+		const additionalContentForNumberOnlydate = [
+			...Array(Math.max(0, groupsToCheck.length - 1)),
+		].map(() => 0);
 
-	const { groups } = matches;
+		return [numberValue, ...additionalContentForNumberOnlydate];
+	}
 
-	const output = groupsToCheck.reduce((accumulator, groupName) => {
-		const value = Number(groups[groupName]);
+	const stringValue = getMetadataKey(cachedMetadata, key, "string");
 
-		// In the case of a faulty regex given by the user in the settings
-		if (!isNaN(value)) accumulator.push(value);
-		return accumulator;
-	}, [] as AbstractDate);
-
-	// Malformed payload bail out
-	if (output.length !== groupsToCheck.length) return undefined;
-
-	return output;
+	if (!stringValue) return undefined;
+	return parseAbstractDate(
+		groupsToCheck,
+		stringValue,
+		settings.dateParserRegex
+	);
 }
